fix(configuration): parseYaml before create configuration
