+++ app/kubernetes/components/kubernetes-configuration-data/kubernetesConfigurationDataController.js
@@ -5,7 +5,6 @@ import { Base64 } from 'js-base64';
 import KubernetesFormValidationHelper from 'Kubernetes/helpers/formValidationHelper';
 import KubernetesConfigurationHelper from 'Kubernetes/helpers/configurationHelper';
 import { KubernetesConfigurationEntry } from 'Kubernetes/models/configuration/models';
-import YAML from 'yaml';
 
 class KubernetesConfigurationDataController {
   /* @ngInject */
@@ -76,27 +75,12 @@ class KubernetesConfigurationDataController {
 
   showSimpleMode() {
     this.formValues.IsSimple = true;
-    YAML.defaultOptions.customTags = ['binary'];
-    const data = YAML.parse(this.formValues.DataYaml);
-    this.formValues.Data = _.map(data, (value, key) => {
-      const entry = new KubernetesConfigurationFormValuesDataEntry();
-      entry.Key = key;
-      entry.Value = value;
-      return entry;
-    });
+    this.formValues.Data = KubernetesConfigurationHelper.parseYaml(this.formValues);
   }
 
   showAdvancedMode() {
     this.formValues.IsSimple = false;
-    const data = _.reduce(
-      this.formValues.Data,
-      (acc, entry) => {
-        acc[entry.Key] = entry.Value;
-        return acc;
-      },
-      {}
-    );
-    this.formValues.DataYaml = YAML.stringify(data);
+    this.formValues.DataYaml = KubernetesConfigurationHelper.parseData(this.formValues);
   }
 
   $onInit() {
